+++ src/Core/Factory.lua
@@ -3,13 +3,29 @@ Contains a list of classes that can be instantiated by the library.
 ]]
 self.classes = {}
 
---[[
+--[[--
 Registers a class so the library is able to instantiate it later.
 
-This method's the same as updating self.classes.
+This method just updates the library classes table by registering a class
+for the client flavors it's supported.
+
+@tparam string classname The name of the class to be registered
+@tparam table classStructure The class structure to be registered
+@tparam nil|string|table clientFlavors The client flavors the class is supported by
 ]]
-function self:addClass(classname, classStructure)
-    self.classes[classname] = classStructure
+function self:addClass(classname, classStructure, clientFlavors)
+    local arr = self.arr
+
+    clientFlavors = arr:wrap(clientFlavors or {
+        self.environment.constants.CLIENT_CLASSIC,
+        self.environment.constants.CLIENT_CLASSIC_ERA,
+        self.environment.constants.CLIENT_RETAIL,
+        self.environment.constants.TEST_SUITE,
+    })
+
+    arr:each(clientFlavors, function(clientFlavor)
+        arr:set(self.classes, clientFlavor .. '.' .. classname, classStructure)
+    end)
 end
 
 --[[
@@ -20,7 +36,9 @@ This method's the same as accessing self.classes[classname].
 @tparam string classname The name of the class to be returned
 ]]
 function self:getClass(classname)
-    return self.classes[classname]
+    local clientFlavor = self.environment:getClientFlavor()
+
+    return self.classes[clientFlavor][classname]
 end
 
 --[[